refact:优化删除业务接口

+++ eva-infra/src/main/java/edu/cuit/infra/gateway/impl/courseimpl/CourseDeleteGatewayImpl.java
@@ -3,6 +3,7 @@ package edu.cuit.infra.gateway.impl.courseimpl;
 import com.baomidou.mybatisplus.core.conditions.Wrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
+import edu.cuit.client.dto.data.course.CoursePeriod;
 import edu.cuit.domain.gateway.course.CourseDeleteGateway;
 import edu.cuit.infra.dal.database.dataobject.course.*;
 import edu.cuit.infra.dal.database.dataobject.eva.EvaTaskDO;
@@ -12,6 +13,8 @@ import edu.cuit.infra.dal.database.mapper.course.*;
 import edu.cuit.infra.dal.database.mapper.eva.EvaTaskMapper;
 import edu.cuit.infra.dal.database.mapper.eva.FormRecordMapper;
 import edu.cuit.infra.dal.database.mapper.user.SysUserMapper;
+import edu.cuit.zhuyimeng.framework.common.exception.QueryException;
+import edu.cuit.zhuyimeng.framework.common.exception.UpdateException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,12 +37,15 @@ public class CourseDeleteGatewayImpl implements CourseDeleteGateway {
 
     @Override
     @Transactional
-    public Void deleteCourses(Integer semId, Integer id, Integer startWeek, Integer endWeek) {
-        //先根据semId和id来找出课程数据
+    public Void deleteCourses(Integer semId, Integer id, CoursePeriod coursePeriod) {
+        //id来找出课程数据
         QueryWrapper<CourInfDO> courseWrapper=new QueryWrapper<>();
         courseWrapper.eq("course_id",id);
-        courseWrapper.between("week",startWeek,endWeek);
-        courInfMapper.delete(courseWrapper);
+        isEmptiy(courseWrapper,coursePeriod);
+        int delete = courInfMapper.delete(courseWrapper);
+        if(delete==0){
+            throw new UpdateException("该节课不存在");
+        }
 
         return null;
     }
@@ -51,8 +57,13 @@ public class CourseDeleteGatewayImpl implements CourseDeleteGateway {
         //删除课程表
         UpdateWrapper<CourseDO> courseWrapper=new UpdateWrapper<>();
         courseWrapper.eq("id",id);
-        courseWrapper.eq("semester_id",semId);
-        courseMapper.delete(courseWrapper);
+       if(semId!=null){
+           courseWrapper.eq("semester_id",semId);
+       }
+        int delete = courseMapper.delete(courseWrapper);
+       if(delete==0){
+           throw new UpdateException("该课程不存在");
+       }
         //删除课程详情表
         UpdateWrapper<CourInfDO> courInfoWrapper=new UpdateWrapper<>();
         courseWrapper.eq("course_id",id);
@@ -81,12 +92,13 @@ public class CourseDeleteGatewayImpl implements CourseDeleteGateway {
     @Override
     @Transactional
     public Void deleteCourseType(List<Integer> ids) {
+        if(ids==null){
+            throw new RuntimeException("请选择要删除的课程类型");
+        }
         QueryWrapper<CourseTypeCourseDO> wrapper = new QueryWrapper<>();
         wrapper.in("type_id",ids);
         //将courseTypeCourse逻辑删除
         courseTypeCourseMapper.delete(wrapper);
-        List<CourseTypeCourseDO> courseTypeCourseDOS = courseTypeCourseMapper.selectList(wrapper);
-
         // 将对应课程类型的逻辑删除
         UpdateWrapper<CourseTypeDO> courseTypeWrapper=new UpdateWrapper<>();
         courseTypeWrapper.in("id",ids);
@@ -98,17 +110,23 @@ public class CourseDeleteGatewayImpl implements CourseDeleteGateway {
     @Override
     @Transactional
     public Void deleteSelfCourse(String userName, Integer courseId) {
+        if(userName==null){
+            throw new QueryException("请先登录");
+        }
         //先根据userName来找到用户id
         Integer userId = userMapper.selectOne(new QueryWrapper<SysUserDO>().eq("username", userName)).getId();
+        if(userId==null){
+            throw new QueryException("你已经被删除了");
+        }
         //根据userId和courseId来删除课程表
         CourseDO courseDO = courseMapper.selectOne(new QueryWrapper<CourseDO>().eq("id", courseId).eq("teacher_id", userId));
         if(courseDO==null){
-            //课程不存在(抛出异常)
-            throw new RuntimeException("没有该用户对应课程");
+            throw new QueryException("没有该用户对应课程");
         }
         courseMapper.delete(new UpdateWrapper<CourseDO>().eq("id", courseId).eq("teacher_id", userId));
         subjectMapper.delete(new UpdateWrapper<SubjectDO>().eq("id", courseDO.getSubjectId()));
         courInfMapper.delete(new UpdateWrapper<CourInfDO>().eq("course_id", courseId));
+        courseTypeCourseMapper.delete(new UpdateWrapper<CourseTypeCourseDO>().eq("course_id", courseId));
         //删除评教相关数据
         List<EvaTaskDO> taskIds = evaTaskMapper.selectList(new QueryWrapper<EvaTaskDO>().eq("cour_inf_id", courseId));
         evaTaskMapper.delete(new UpdateWrapper<EvaTaskDO>().eq("cour_inf_id", courseId));
@@ -116,5 +134,23 @@ public class CourseDeleteGatewayImpl implements CourseDeleteGateway {
 
 
         return null;
+    }
+
+    private void isEmptiy(QueryWrapper wrapper,CoursePeriod coursePeriod){
+            if(coursePeriod.getStartWeek()!=null){
+                wrapper.ge("week", coursePeriod.getStartWeek());
+            }
+            if(coursePeriod.getEndWeek()!=null){
+                wrapper.le("week", coursePeriod.getEndWeek());
+            }
+            if(coursePeriod.getDay()!=null){
+                wrapper.eq("day", coursePeriod.getDay());
+            }
+            if(coursePeriod.getStartTime()!=null){
+                wrapper.eq("start_time", coursePeriod.getStartTime());
+            }
+        if(coursePeriod.getEndTime()!=null){
+            wrapper.eq("end_time", coursePeriod.getEndTime());
+        }
     }
 }
